feat: rename icon with all the primary icons to "IconPrimary" resp. "icon-primary"
