+++ packages/dnb-design-system-portal/src/pages/icons/primary.md
@@ -11,7 +11,7 @@ import Icons from "Parts/icons/ListAllIcons";
 
 The Primary Icons are included in the [UI Library](/uilib) and also shipped bundled as an [UMD Package](https://unpkg.com/dnb-ui-lib@latest/umd/dnb-ui-lib-icons.min.js).
 
-Primary Icons an easily be included, for instance, in [Buttons](/uilib/components/button). They are integrated by using the `<IconWithAllIcons />` Component.
+Primary Icons an easily be included, for instance, in [Buttons](/uilib/components/button). They are integrated by using the `<IconPrimary />` Component.
 
 #### A list of all Primary Icons
