+++ packages/dnb-design-system-portal/src/uilib/components/examples/Icon.txt
@@ -1,3 +1,3 @@
-<IconWithAllIcons icon="question" />
-<IconWithAllIcons icon="chevron_right_medium" />
+<IconPrimary icon="question" />
+<IconPrimary icon="chevron_right_medium" />
 <Icon icon={Bell} size="32" />
