+++ packages/dnb-design-system-portal/src/uilib/scripts/makeDemosFactory.js
@@ -10,11 +10,7 @@ import runFactory from './tasks/makerFactory'
 import processAllPartsFile from './tasks/processAllPartsFile'
 
 if (require.main === module && process.env.NODE_ENV !== 'test') {
-  const ignoreUiLibList = [
-    'web-components',
-    'style',
-    'icon-with-all-icons'
-  ]
+  const ignoreUiLibList = ['web-components', 'style', 'icon-primary']
   const keepFiles = [] //'!**/footer/*'
   const autoAdvice = `
   ATTENTION: This file is auto generated by using "makeDemosFactory".
