+++ packages/dnb-ui-lib/src/components/dropdown/Dropdown.js
@@ -14,7 +14,7 @@ import {
   dispatchCustomElementEvent
 } from '../../shared/component-helper'
 // import './style/dnb-dropdown.scss' // no good solution to import the style here
-import Icon from '../icon/IconWithAllIcons'
+import Icon from '../icon/IconPrimary'
 
 const renderProps = {
   on_show: null,
