+++ packages/dnb-ui-lib/src/components/icon-primary/description.md
@@ -1,0 +1,1 @@
+TODO: Write a description for "icon-primary"
