+++ packages/dnb-ui-lib/src/components/icon-primary/index.js
@@ -1,0 +1,6 @@
+/**
+ * Button Component
+ *
+ */
+
+export * from './IconPrimary'
