+++ packages/dnb-ui-lib/src/components/icon-primary/style.js
