+++ packages/dnb-ui-lib/src/components/icon-primary/web-component.js
