--- packages/dnb-ui-lib/src/components/icon-with-all-icons/description.md
@@ -1,1 +1,0 @@
-TODO: Write a description for "icon-with-all-icons"
