--- packages/dnb-ui-lib/src/components/icon-with-all-icons/index.js
@@ -1,6 +1,0 @@
-/**
- * Button Component
- *
- */
-
-export * from './IconWithAllIcons'
