--- packages/dnb-ui-lib/src/components/icon-with-all-icons/style/_icon-with-all-icons.scss
@@ -1,1 +1,0 @@
-/* Dummy File */
