+++ packages/dnb-ui-lib/src/components/icon/Example.js
@@ -5,7 +5,7 @@
 
 import React, { PureComponent, Fragment } from 'react'
 import Icon from './Icon'
-import IconWithAllIcons from './IconWithAllIcons'
+import IconPrimary from './IconPrimary'
 import { bell_medium as Bell } from './lib'
 
 class Example extends PureComponent {
@@ -26,8 +26,8 @@ import bell from './icons/bell'
   render() {
     return (
       <Fragment>
-        <IconWithAllIcons icon="question" />
-        <IconWithAllIcons icon="chevron_right_medium" />
+        <IconPrimary icon="question" />
+        <IconPrimary icon="chevron_right_medium" />
         <Icon icon={Bell} size="32" />
       </Fragment>
     )
