+++ packages/dnb-ui-lib/src/components/icon/description.md
@@ -2,4 +2,4 @@ The main Icon component is basically only a wrapper for what ever Icon You send 
 
 #### More details
 
-To make it clear that all the "often used" icons are loaded at once, we have an additional Icon component, named `IconWithAllIcons` or `dnb-icon-with-all-icons`.
+To make it clear that all the "often used" icons are loaded at once, we have an additional Icon component, named `IconPrimary` or `dnb-icon-primary`.
