+++ packages/dnb-ui-lib/src/components/icon/details.md
@@ -1,6 +1,6 @@
 | Properties   | Description                                                                                                                                                      |
 | ------------ | ---------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `icon`       | _(mandatory)_ a React SVG Component or the icon name (in case we use `IconWithAllIcons` or `dnb-icon-with-all-icons`).                                           |
+| `icon`       | _(mandatory)_ a React SVG Component or the icon name (in case we use `IconPrimary` or `dnb-icon-primary`).                                                       |
 | `alt`        | _(mandatory)_ the alternative label (text version) of the icon.                                                                                                  |
 | `size`       | _(optional)_ the dimension of the icon. This will be the `viewBox` and represent `width` and `height`. Defaults to `16`. You can use `small`,`medium` or `large` |
 | `color`      | _(optional)_ sets a color property to the `svg` markup. Default is no color, witch means _black_                                                                 |
