+++ packages/dnb-ui-lib/src/index.js
@@ -18,7 +18,7 @@ import Button from './components/button/Button'
 import Dropdown from './components/dropdown/Dropdown'
 import FormLabel from './components/form-label/FormLabel'
 import Icon from './components/icon/Icon'
-import IconWithAllIcons from './components/icon-with-all-icons/IconWithAllIcons'
+import IconPrimary from './components/icon-primary/IconPrimary'
 import Input from './components/input/Input'
 import InputMasked from './components/input-masked/InputMasked'
 import LineTitle from './components/line-title/LineTitle'
@@ -50,7 +50,7 @@ export {
   Dropdown,
   FormLabel,
   Icon,
-  IconWithAllIcons,
+  IconPrimary,
   Input,
   InputMasked,
   LineTitle,
