+++ packages/dnb-ui-lib/src/style/dnb-ui-components.scss
@@ -9,7 +9,6 @@
 @import '../components/dropdown/style/_dropdown.scss';
 @import '../components/form-label/style/_form-label.scss';
 @import '../components/icon/style/_icon.scss';
-@import '../components/icon-with-all-icons/style/_icon-with-all-icons.scss';
 @import '../components/input/style/_input.scss';
 @import '../components/input-masked/style/_input-masked.scss';
 @import '../components/line-title/style/_line-title.scss';
