+++ packages/examples/example-html/static/umd/index.html
@@ -13,7 +13,7 @@
   <dnb-button text="Declarative" icon="chevron-right" on_click="myClass.on_click"></dnb-button>
   <dnb-button text="Imperative" icon="chevron-right" class="button"></dnb-button>
 
-  <dnb-icon-with-all-icons icon="bell" size="80"></dnb-icon-with-all-icons>
+  <dnb-icon-primary icon="bell" size="80"></dnb-icon-primary>
 
   <dnb-form-label for_id="switch-1">Switch</dnb-form-label>
   <dnb-switch id="switch-1" title="Switch title"></dnb-switch>
